refactor(epic4): separate planning runtime from routes
Redirect the root route to Home and move planning composition out of the app route layer. Keep runtime construction isolated from planning UI for later screen splits.
