+++ src/app/list/[id].tsx
@@ -1,6 +1,6 @@
 import { useLocalSearchParams } from 'expo-router';
 
-import HomeScreen from '../index';
+import HomeScreen from '@/components/planning/home-screen';
 
 export default function ListDetailRoute() {
   const { id } = useLocalSearchParams<{ id: string }>();
