+++ src/app/list/new.tsx
@@ -1,4 +1,4 @@
-import HomeScreen from '../index';
+import HomeScreen from '@/components/planning/home-screen';
 
 export default function CreateListRoute() {
   return <HomeScreen routeIntent={{ kind: 'new-list' }} />;
